requirements.txt
boto3
pandas
flask
gunicorn
openpyxl
xlsxwriter
python-dotenv